backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
pyjwt==2.8.0
gunicorn==22.0.0
sqlalchemy==2.0.31